+++ tensorflow/compiler/tf2xla/xla_helpers.cc
@@ -190,7 +190,7 @@ StatusOr<xla::XlaOp> ReshapeWithCorrectRepresentationAndSharding(
     absl::optional<xla::OpSharding> sharding, bool fast_mem) {
   if (original_shape.IsTuple()) {
     std::vector<xla::XlaOp> elements;
-    for (int64_t i = 0; i < original_shape.tuple_shapes_size(); ++i) {
+    for (int i = 0; i < original_shape.tuple_shapes_size(); ++i) {
       auto subsharding = sharding ? sharding->tuple_shardings(i) : sharding;
       TF_ASSIGN_OR_RETURN(auto element,
                           ReshapeWithCorrectRepresentationAndSharding(
